update auto-organisation pipeline to support move history

+++ app/src/main/java/com/fpf/smartscan/data/moveHistory/Dao.kt
@@ -6,12 +6,12 @@ import androidx.room.*
 interface MoveHistoryDao {
 
     @Query("SELECT * FROM move_history  WHERE scanId = :scanId ORDER BY id DESC")
-    fun getMoveHistory(scanId: Long): List<MoveHistory>
+    fun getMoveHistory(scanId: Int): List<MoveHistory>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertMove(moveHistory: MoveHistory)
 
     @Query("DELETE FROM move_history WHERE scanId = :scanId")
-    suspend fun deleteMoveHistory(scanId: Long)
+    suspend fun deleteMoveHistory(scanId: Int)
 }
